+++ psengine/playbook_alerts/__init__.py
@@ -28,6 +28,7 @@ from .playbook_alerts import (
     PBA_CyberVulnerability,
     PBA_DomainAbuse,
     PBA_Generic,
+    PBA_GeopoliticsFacility,
     PBA_IdentityNovelExposure,
     PBA_MalwareReport,
     PBA_ThirdPartyRisk,