+++ docs/CHANGELOG.md
@@ -1,10 +1,29 @@
 # Changelog
 
+## v2.2.0 - 2025-10-08
+
+### Added
+
+- `DetectionMgr.search` supports relative dates for `created_` and `updated_` fields.
+- `ClassicAlert.markdown` does not require `triggered_by` field to be retrieved when not used in markdown.
+- `LookupMgr.lookup` supports `organization` as entity type.
+- `EnrichmentData` now has a `links` method to facilitate the extraction of links of the needed type from a specific section.
+
+### Fixed
+
+- `playbook_alert.PBA_GeopoliticsFacility` can now be imported from the module level.
+- `playbook_alert.PBA_GeopoliticsFacility.markdown` properly shows alert events.
+
+### Changed
+
+- `SoarMgr.SOAR_POST_ROWS` decreased to `1000` to comply with Product guidelines.
+- `SoarMgr.soar` uses the new v3 SOAR API endpoint.
+
 ## v2.1.1 - 2025-08-22
 
 ### Changed
 
-- Moved typing system to `typing.Annotated` and `typing_extensions.Doc`
+- Moved typing system to `typing.Annotated` and `typing_extensions.Doc`.
 
 ## v2.1.0 - 2025-07-02